ref(handlers): split handlers into config, key
- split handlers into top level commands config,
key and descriptors
